--- components/src/main/java/ru/s4m1d/notes/app/desktop/client/components/constant/ComponentName.java
@@ -1,6 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.components.constant;
-
-public class ComponentName {
-    public static final String MAIN_FRAME = "MainFrame";
-    public static final String NOTES_PANE = "NotesPane";
-}
