--- components/src/main/java/ru/s4m1d/notes/app/desktop/client/components/layout/NoInsetsScrollPaneLayout.java
@@ -1,22 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.components.layout;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class NoInsetsScrollPaneLayout extends ScrollPaneLayout {
-//    @Override
-//    public void layoutContainer(Container parent) {
-//        JScrollPane scrollPane = (JScrollPane) parent;
-//        Rectangle availableRectangle = scrollPane.getBounds();
-//        Rectangle vsbRectangle = new Rectangle();
-//        vsbRectangle.height = scrollPane.getHeight();
-//        vsbRectangle.width = 12;
-//        vsbRectangle.x = availableRectangle.x + availableRectangle.width - vsbRectangle.width;
-//        vsbRectangle.y = availableRectangle.y;
-//        vsb.setVisible(true);
-//        vsb.setBounds(vsbRectangle);
-//
-//        viewport.setBounds(availableRectangle);
-//
-//    }
-}
