--- components/src/main/java/ru/s4m1d/notes/app/desktop/client/components/layout/StrategyNotFoundException.java
@@ -1,4 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.components.layout;
-
-public class StrategyNotFoundException extends Exception{
-}
