--- components/src/main/java/ru/s4m1d/notes/app/desktop/client/components/layout/strategy/LayeredPaneLayoutStrategy.java
@@ -1,7 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.components.layout.strategy;
-
-import java.awt.*;
-
-public interface LayeredPaneLayoutStrategy <T extends Component> {
-    void layoutComponent(T component);
-}
