--- components/src/main/java/ru/s4m1d/notes/app/desktop/client/components/layout/strategy/ScrollPaneLayoutStrategy.java
@@ -1,12 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.components.layout.strategy;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class ScrollPaneLayoutStrategy implements LayeredPaneLayoutStrategy<JScrollPane>{
-    @Override
-    public void layoutComponent(JScrollPane component) {
-        Rectangle availableRectangle = component.getParent().getBounds();
-        component.setBounds(availableRectangle);
-    }
-}
