+++ presenter/pom.xml
@@ -19,7 +19,7 @@
     <dependencies>
         <dependency>
             <groupId>ru.s4m1d</groupId>
-            <artifactId>components</artifactId>
+            <artifactId>view</artifactId>
             <version>${parent.version}</version>
         </dependency>
         <dependency>
