+++ presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/mainframe/MainFrameStateService.java
@@ -1,6 +1,6 @@
 package ru.s4m1d.notes.app.desktop.client.presenter.mainframe;
 
-import ru.s4m1d.notes.app.desktop.client.components.MainFrame;
+import ru.s4m1d.notes.app.desktop.client.view.MainFrame;
 import ru.s4m1d.notes.app.desktop.client.core.Observer;
 
 public class MainFrameStateService extends Observer<NewComponentEvent> {
