+++ view/pom.xml
@@ -9,29 +9,11 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>app</artifactId>
+    <artifactId>view</artifactId>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
     </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>ru.s4m1d</groupId>
-            <artifactId>view</artifactId>
-            <version>${parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>ru.s4m1d</groupId>
-            <artifactId>daemon-impl</artifactId>
-            <version>${parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>ru.s4m1d</groupId>
-            <artifactId>presenter</artifactId>
-            <version>${parent.version}</version>
-        </dependency>
-    </dependencies>
 
 </project>
