+++ view/src/main/java/ru/s4m1d/notes/app/desktop/client/view/constant/ComponentName.java
@@ -1,0 +1,6 @@
+package ru.s4m1d.notes.app.desktop.client.view.constant;
+
+public class ComponentName {
+    public static final String MAIN_FRAME = "MainFrame";
+    public static final String NOTES_PANE = "NotesPane";
+}
