+++ view/src/main/java/ru/s4m1d/notes/app/desktop/client/view/layout/StrategyNotFoundException.java
@@ -1,0 +1,4 @@
+package ru.s4m1d.notes.app.desktop.client.view.layout;
+
+public class StrategyNotFoundException extends Exception{
+}
