+++ view/src/main/java/ru/s4m1d/notes/app/desktop/client/view/layout/strategy/LayeredPaneLayoutStrategy.java
@@ -1,0 +1,7 @@
+package ru.s4m1d.notes.app.desktop.client.view.layout.strategy;
+
+import java.awt.*;
+
+public interface LayeredPaneLayoutStrategy <T extends Component> {
+    void layoutComponent(T component);
+}
